+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 0.2.0 - TBA
 -----------
 
+- Ability to remove headers from email message. `#24`_
 - Improved attachments interface. `#18`_
 - Support for sending single email. `#11`_
 
@@ -20,5 +21,6 @@ Changelog
 - Initial release.
 
 
+.. _#24: https://github.com/FriendlyCoders/postmarker/issues/24
 .. _#18: https://github.com/FriendlyCoders/postmarker/issues/18
 .. _#11: https://github.com/FriendlyCoders/postmarker/issues/11